fix(cran): resolve Windows link errors in stub layer
- Stub files declared Praat APIs with 'long' where Praat uses 'integer'
  (intptr_t). On LP64 macOS both mangle identically so it linked; on
  Windows LLP64 'integer' is 'long long', so the stubs mangled differently
  and the DLL link failed with ~20 undefined Ui*/Editor/Graphics/audio
  symbols. Replace bare 'long' with 'integer' throughout the stub
  signatures (uiform_stubs, sound_audio_stubs,
  graphics_stubs_comprehensive, praat_stubs, praat_app_stubs).
- Add FunctionEditor_init stub: never had one; the macOS linker tolerated
  the dangling reference, the Windows linker does not.
- Makevars.win: link -lgdiplus (Praat Photo/GraphicsScreen use the GDI+
  flat API behind _WIN32 guards).
- Makevars: regenerated by configure with the libc++ deprecation define
  (kept in sync with Makevars.in).

Verified: mingw cross-compile of stubs + nm confirms every previously
undefined symbol now resolves; macOS install green with bit-identical
formant output.

+++ src/graphics_stubs_comprehensive.cpp
@@ -200,7 +200,7 @@ void Sound_draw_btlr (Sound me, Graphics g, double xmin, double xmax, double ymi
     Melder_throw (U"Sound_draw_btlr: Graphics functions are not available in this build.");
 }
 
-void Graphics_WCtoDC (Graphics, double, double, long *, long *) { /* No-op */ }
+void Graphics_WCtoDC (Graphics, double, double, integer *, integer *) { /* No-op */ }
 
 // Screen graphics creation (for GUI)
 autoGraphics Graphics_create_screen (void *, void *, int) {
